requirements.txt
numpy>=1.17
scipy>=1.5
matplotlib>=3.2
biopython>=1.7
